% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_discreteRV.R
\name{discreteRV}
\alias{discreteRV}
\title{Discrete Random Variables}
\usage{
discreteRV(f, type = "PMF", lower = 0, upper = Inf, support = NULL)
}
\arguments{
\item{f}{A function, where f(x) returns the value of the defining function at x.}

\item{type}{The type of function that the defining function is ("PMF", "CDF", etc.).
Defaults to "PMF".}

\item{lower}{The lower limit of the support of the distribution. Defaults to 0.}

\item{upper}{The upper limit of the support of the distribution. Defaults to Inf.}

\item{support}{A vector of numeric or integer values where the PMF is positive.
If support is not NULL, upper and lower will be ignored. If support is NULL,
upper and lower will be used to define the support. Defaults to NULL.}
}
\value{
An object of class "discreteRV".
}
\description{
This function initializes a discrete random variable object using a defining
  function, an indicator of what type the defining function is ("PMF", "CDF", etc.),
  and a description of the support. The support can be given either as a vector of
  values or as upper and lower limits. If upper and lower limits are supplied,
  the support is assumed to be all integers between the upper and lower limits,
  inclusive. Currently, this supports univariate, discrete random variables.
}
