% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{expectedValue.discreteRV}
\alias{expectedValue.discreteRV}
\title{Calculate Expected Value of A Discrete Random Variable}
\usage{
\method{expectedValue}{discreteRV}(X)
}
\arguments{
\item{X}{An object of class "discreteRV".}
}
\value{
A length one numeric, the expected value of the random variable X.
}
\description{
This function calculates the expected value of a discrete random variable.
}
