% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{expectedValue}
\alias{expectedValue}
\title{Calculate Expected Value}
\usage{
expectedValue(X, ...)
}
\arguments{
\item{X}{A random variable object.}
}
\value{
The expected value.
}
\description{
This function calculates the expected value of an object.
}
