% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{expectedValue.RV}
\alias{expectedValue.RV}
\title{Calculate Expected Value of A Random Variable}
\usage{
\method{expectedValue}{RV}(X)
}
\arguments{
\item{X}{An object of class "RV".}
}
\value{
A length one numeric, the expected value of the random variable X.
}
\description{
This function calculates the expected value of a random variable.
}
