% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_variates.R
\name{generateVariates}
\alias{generateVariates}
\title{Generic Method to Generate Random Variates}
\usage{
generateVariates(X, ...)
}
\arguments{
\item{X}{An object.}
}
\value{
An object.
}
\description{
This function generates random variates from an object.
}
