% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions_discrete.R
\name{getCDF.discreteRV}
\alias{getCDF.discreteRV}
\title{Get the CDF from a Discrete Random Variable Object}
\usage{
\method{getCDF}{discreteRV}(X)
}
\arguments{
\item{X}{An object of class "discreteRV".}
}
\value{
A function that evaluates the CDF of X.
}
\description{
This function extracts the CDF from a discrete random variable object.
}
