% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getCDF}
\alias{getCDF}
\title{Generic Method to Get a CDF}
\usage{
getCDF(X, ...)
}
\arguments{
\item{X}{An object.}
}
\value{
An object.
}
\description{
This function extracts the CDF from an object.
}
