% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getCDF.RV}
\alias{getCDF.RV}
\title{Get the CDF from a Random Variable Object}
\usage{
\method{getCDF}{RV}(X)
}
\arguments{
\item{X}{An object of class "RV".}
}
\value{
A function that evaluates the CDF of X.
}
\description{
This function extracts the CDF from a random variable object.
}
