% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions_discrete.R
\name{getCHF.discreteRV}
\alias{getCHF.discreteRV}
\title{Get the Cumulative Hazard Function from a Discrete Random Variable Object}
\usage{
\method{getCHF}{discreteRV}(X)
}
\arguments{
\item{X}{An object of class "discreteRV".}
}
\value{
A function that evaluates the CHF of X.
}
\description{
This function extracts the CHF from a discrete random variable object.
}
