% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getCHF}
\alias{getCHF}
\title{Get a cumulative hazard function}
\usage{
getCHF(X, ...)
}
\arguments{
\item{X}{An object for which a cumulative hazard function can be defined.}
}
\value{
A function.
}
\description{
This is a generic function to get the cumulative hazard function from
  an object.
}
