% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getCHF.RV}
\alias{getCHF.RV}
\title{Get the Cumulative Hazard Function from a Random Variable Object}
\usage{
\method{getCHF}{RV}(X)
}
\arguments{
\item{X}{An object of class "RV".}
}
\value{
A function that evaluates the CHF of X.
}
\description{
This function extracts the CHF from a random variable object.
}
