% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions_discrete.R
\name{getHF.discreteRV}
\alias{getHF.discreteRV}
\title{Get the Hazard Function from a Discrete Random Variable Object}
\usage{
\method{getHF}{discreteRV}(X)
}
\arguments{
\item{X}{An object of class "discreteRV".}
}
\value{
A function that evaluates the HF of X.
}
\description{
This function extracts the HF from a discrete random variable object.
}
