% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getHF}
\alias{getHF}
\title{Generic Method to Get a Hazard Function}
\usage{
getHF(X, ...)
}
\arguments{
\item{X}{An object.}
}
\value{
An object.
}
\description{
This function extracts the hazard function from an object.
}
