% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getHF.RV}
\alias{getHF.RV}
\title{Get the Hazard Function from a Random Variable Object}
\usage{
\method{getHF}{RV}(X)
}
\arguments{
\item{X}{An object of class "RV".}
}
\value{
A function that evaluates the HF of X.
}
\description{
This function extracts the HF from a random variable object.
}
