% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions_discrete.R
\name{getInverseCDF.discreteRV}
\alias{getInverseCDF.discreteRV}
\title{Get the Inverse CDF from a Discrete Random Variable Object}
\usage{
\method{getInverseCDF}{discreteRV}(X)
}
\arguments{
\item{X}{An object of class "discreteRV".}
}
\value{
A function that evaluates the Inverse CDF of X.
}
\description{
This function extracts the Inverse CDF from a discrete random variable object.
}
