% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getInverseCDF}
\alias{getInverseCDF}
\title{Generic Method to Get an Inverse CDF}
\usage{
getInverseCDF(X, ...)
}
\arguments{
\item{X}{An object.}
}
\value{
An object.
}
\description{
This function extracts the inverse CDF from an object.
}
