% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getInverseCDF.RV}
\alias{getInverseCDF.RV}
\title{Get the Inverse CDF from a Random Variable Object}
\usage{
\method{getInverseCDF}{RV}(X)
}
\arguments{
\item{X}{An object of class "RV".}
}
\value{
A function that evaluates the Inverse CDF of X.
}
\description{
This function extracts the Inverse CDF from a random variable object.
}
