% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions_discrete.R
\name{getMGF.discreteRV}
\alias{getMGF.discreteRV}
\title{Get the Moment Generating Function from a Discrete Random Variable Object}
\usage{
\method{getMGF}{discreteRV}(X)
}
\arguments{
\item{X}{An object of class "discreteRV".}
}
\value{
A function that evaluates the MGF of X.
}
\description{
This function extracts the MGF from a discrete random variable object.
}
