% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getMGF}
\alias{getMGF}
\title{Get a moment generating function}
\usage{
getMGF(X, ...)
}
\arguments{
\item{X}{An object for which a moment generating function can be defined.}
}
\value{
A function.
}
\description{
This is a generic function to get the moment generating function from
  an object.
}
