% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getMGF.RV}
\alias{getMGF.RV}
\title{Get the Moment Generating Function from a Random Variable Object}
\usage{
\method{getMGF}{RV}(X)
}
\arguments{
\item{X}{An object of class "RV".}
}
\value{
A function that evaluates the MGF of X.
}
\description{
This function extracts the MGF from a random variable object.
}
