% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getPDF}
\alias{getPDF}
\title{Generic Method to Get a PDF}
\usage{
getPDF(X, ...)
}
\arguments{
\item{X}{An object.}
}
\value{
An object.
}
\description{
This function extracts the PDF from an object.
}
