% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getPDF.RV}
\alias{getPDF.RV}
\title{Get the PDF from a Random Variable Object}
\usage{
\method{getPDF}{RV}(X)
}
\arguments{
\item{X}{An object of class "RV".}
}
\value{
A function that evaluates the PDF of X.
}
\description{
This function extracts the PDF from a random variable object.
}
