% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions_discrete.R
\name{getPMF.discreteRV}
\alias{getPMF.discreteRV}
\title{Get the PMF from a Discrete Random Variable Object}
\usage{
\method{getPMF}{discreteRV}(X)
}
\arguments{
\item{X}{An object of class "discreteRV".}
}
\value{
A function that evaluates the PMF of X.
}
\description{
This function extracts the PMF from a discrete random variable object.
}
