% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions_discrete.R
\name{getPMF}
\alias{getPMF}
\title{Get a probability mass function}
\usage{
getPMF(X, ...)
}
\arguments{
\item{X}{An object for which a probability mass function can be defined.}
}
\value{
A function.
}
\description{
This is a generic function to get the probability mass function from
  an object.
}
