% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions_discrete.R
\name{getSF.discreteRV}
\alias{getSF.discreteRV}
\title{Get the Survival Function from a Discrete Random Variable Object}
\usage{
\method{getSF}{discreteRV}(X)
}
\arguments{
\item{X}{An object of class "discreteRV".}
}
\value{
A function that evaluates the SF of X.
}
\description{
This function extracts the SF from a discrete random variable object.
}
