% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getSF}
\alias{getSF}
\title{Generic Method to Get a Survival Function}
\usage{
getSF(X, ...)
}
\arguments{
\item{X}{An object.}
}
\value{
An object.
}
\description{
This function extracts the survival function from an object.
}
