% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getSF.RV}
\alias{getSF.RV}
\title{Get the Survival Function from a Random Variable Object}
\usage{
\method{getSF}{RV}(X)
}
\arguments{
\item{X}{An object of class "RV".}
}
\value{
A function that evaluates the SF of X.
}
\description{
This function extracts the SF from a random variable object.
}
