% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_discrete_rvs.R
\name{\%convolution\%.discreteRV}
\alias{\%convolution\%.discreteRV}
\title{Find the Distribution of a Convolution of Discrete Random Variables}
\usage{
`\%convolution\%.discreteRV`(X, Y)
}
\arguments{
\item{X}{An object of class "discreteRV".}

\item{Y}{An object of class "discreteRV".}
}
\value{
An object of class "discreteRV", whose distribution is that of X + Y.
}
\description{
This function finds the distribution of a convolution of two discrete
  random variables. The two discrete random variables are assumed to be
  independent. The result is returned as an object of class "discreteRV".
}
