% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_rvs.R
\name{\%convolution\%.RV}
\alias{\%convolution\%.RV}
\title{Find the Distribution of a Convolution of Random Variables}
\usage{
`\%convolution\%.RV`(X, Y)
}
\arguments{
\item{X}{An object of class "RV".}

\item{Y}{An object of class "RV".}
}
\value{
An object of class "RV", whose distribution is that of X + Y.
}
\description{
This function finds the distribution of a convolution of two random variables.
  The two random variables are assumed to be independent. The result is
  returned as an object of class "RV".
}
