% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_rvs.R
\name{\%convolution\%}
\alias{\%convolution\%}
\title{Find the Distribution of a Convolution}
\usage{
X \%convolution\% Y
}
\arguments{
\item{X}{An object with a distribution.}

\item{Y}{An object with a distribution.}
}
\value{
An object of the same class as X, whose distribution is
  that of X + Y.
}
\description{
This is a generic function to find the distribution of the convolution of
  two objects.
}
