% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_discrete_rvs.R
\name{mapRV.discreteRV}
\alias{mapRV.discreteRV}
\title{Find the Distribution of a Function of a Discrete Random Variable}
\usage{
\method{mapRV}{discreteRV}(X, g)
}
\arguments{
\item{X}{An object of class "discreteRV".}

\item{g}{A differentiable function.}
}
\value{
An object of class "discreteRV", whose distribution is that of g(X).
}
\description{
This function finds the distribution of a differentiable function of a
  discrete random variable and returns the result as a discreteRV object.
  This function is still experimental.
}
