% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_rvs.R
\name{mapRV}
\alias{mapRV}
\title{Generic Method to Map a Random Variable}
\usage{
mapRV(X, ...)
}
\arguments{
\item{X}{An object.}
}
\value{
An object.
}
\description{
This function finds the distribution of a function of a random variable.
}
