% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_rvs.R
\name{mapRV.RV}
\alias{mapRV.RV}
\title{Find the Distribution of a Function of a Random Variable}
\usage{
\method{mapRV}{RV}(X, g)
}
\arguments{
\item{X}{An object of class "RV".}

\item{g}{A differentiable function.}
}
\value{
An object of class "RV", whose distribution is that of g(X).
}
\description{
This function uses the distribution function technique to find the distribution
  of a differentiable function of a random variable and returns the result
  as an RV object.
}
