% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_discreteRV.R
\name{new_discreteRV}
\alias{new_discreteRV}
\title{Internal Constructor for Discrete Random Variables}
\usage{
new_discreteRV(X = list())
}
\arguments{
\item{X}{A list, to be used to construct a discrete random variable.}
}
\value{
An object of class "discreteRV".
}
\description{
This is the internal construction function for objects of class "discreteRV".
}
