% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rv.R
\name{new_RV}
\alias{new_RV}
\title{Internal Constructor for Random Variables}
\usage{
new_RV(X = list())
}
\arguments{
\item{X}{A list, to be used to construct a random variable.}
}
\value{
An object of class "RV".
}
\description{
This is the internal construction function for objects of class "RV".
}
