% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rv.R
\name{RV}
\alias{RV}
\title{Random Variables}
\usage{
RV(f, type = "PDF", lower = -Inf, upper = Inf)
}
\arguments{
\item{f}{A function, where f(x) returns the value of the defining function at x.}

\item{type}{The type of function that the defining function is ("PDF", "CDF", etc.).}

\item{lower}{The lower limit of the support of the distribution. Defaults to -Inf.}

\item{upper}{The upper limit of the support of the distribution. Defaults to Inf.}
}
\value{
An object of class "RV".
}
\description{
This function initializes a random variable object using a defining function,
  an indicator of what type the defining function is ("PDF", "CDF", etc.),
  and upper and lower limits of the support. Currently, this supports
  univariate, continuous random variables where the support is an interval.
}
