% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_discreteRV.R
\name{validate_discreteRV}
\alias{validate_discreteRV}
\title{Validate Discrete Random Variable Objects}
\usage{
validate_discreteRV(X)
}
\arguments{
\item{X}{An object of class "discreteRV".}
}
\value{
X. Primarily called for the side effect of producing an error message
  if the object is not a valid discrete random variable.
}
\description{
This function checks whether a discrete random variable object is valid.
}
