% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rv.R
\name{validate_RV}
\alias{validate_RV}
\title{Validate Random Variable Objects}
\usage{
validate_RV(X)
}
\arguments{
\item{X}{An object of class "RV".}
}
\value{
X. Primarily called for the side effect of producing an error message
  if the object is not a valid random variable.
}
\description{
This function checks whether a random variable object is valid.
}
