% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{variance.discreteRV}
\alias{variance.discreteRV}
\title{Calculate the Variance of A Discrete Random Variable}
\usage{
\method{variance}{discreteRV}(X)
}
\arguments{
\item{X}{An object of class "discreteRV".}
}
\value{
A length one numeric, the variance of the random variable X.
}
\description{
This function calculates the variance of a discrete random variable.
}
