% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{variance.RV}
\alias{variance.RV}
\title{Calculate the Variance of A Random Variable}
\usage{
\method{variance}{RV}(X)
}
\arguments{
\item{X}{An object of class "RV".}
}
\value{
A length one numeric, the variance of the random variable X.
}
\description{
This function calculates the variance of a random variable.
}
