% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_variates.R
\name{generateVariates}
\alias{generateVariates}
\title{Generate Random Variates}
\usage{
generateVariates(X, ...)
}
\arguments{
\item{X}{A random variable object.}
}
\value{
A numeric vector of random variates.
}
\description{
This is a generic function to generate random variates from a distribution.
}
