% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getCDF}
\alias{getCDF}
\title{Get a cumulative distribution function}
\usage{
getCDF(X, ...)
}
\arguments{
\item{X}{An object for which a cumulative distribution function can be defined.}
}
\value{
A function.
}
\description{
This is a generic function to get the cumulative distribution function from
  an object.
}
