% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getCHF}
\alias{getCHF}
\title{Generic Method to Get a Cumulative Hazard Function}
\usage{
getCHF(X, ...)
}
\arguments{
\item{X}{An object.}
}
\value{
An object.
}
\description{
This function extracts the cumulative hazard function from an object.
}
