% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getHF}
\alias{getHF}
\title{Get a hazard function}
\usage{
getHF(X, ...)
}
\arguments{
\item{X}{An object for which a hazard function can be defined.}
}
\value{
A function.
}
\description{
This is a generic function to get the hazard function from an object.
}
