% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getInverseCDF}
\alias{getInverseCDF}
\title{Get an inverse cumulative distribution function}
\usage{
getInverseCDF(X, ...)
}
\arguments{
\item{X}{An object for which an inverse cumulative distribution function can be defined.}
}
\value{
A function.
}
\description{
This is a generic function to get the inverse cumulative distribution
  function from an object.
}
