% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getMGF}
\alias{getMGF}
\title{Generic Method to Get a Moment Generating Function}
\usage{
getMGF(X, ...)
}
\arguments{
\item{X}{An object.}
}
\value{
An object.
}
\description{
This function extracts the moment generating function from an object.
}
