% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getPDF}
\alias{getPDF}
\title{Get a probability density function}
\usage{
getPDF(X, ...)
}
\arguments{
\item{X}{An object for which a probability density function can be defined.}
}
\value{
A function.
}
\description{
This is a generic function to get the probability density function from
  an object.
}
