% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functions.R
\name{getSF}
\alias{getSF}
\title{Get a survival function}
\usage{
getSF(X, ...)
}
\arguments{
\item{X}{An object for which a survival function can be defined.}
}
\value{
A function.
}
\description{
This is a generic function to get the survival function from an object.
}
