% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_of_rvs.R
\name{mapRV}
\alias{mapRV}
\title{Apply a Function to a Random Variable}
\usage{
mapRV(X, ...)
}
\arguments{
\item{X}{A random variable object.}
}
\value{
Another random variable object.
}
\description{
This is a generic function to find the distribution of a function
  of a random variable.
}
