% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_metrics.R
\name{variance}
\alias{variance}
\title{Calculate Variance}
\usage{
variance(X, ...)
}
\arguments{
\item{X}{An object.}
}
\value{
An object.
}
\description{
This function calculates the variance of an object.
}
